moved health related commands to their own plugin
This is part 4 of #365 after #402.
